LPS-168928 feature-flag-web: removes FeatureFlagPropsManager
